initial EditDialog works
EditDialog: refactor EditContext to work with DataItem not Feature
